switch to long double

+++ PSiVR/PSiVR/Geometry/mat3.cpp
@@ -1,18 +1,18 @@
 #include "mat3.h"
 
-const mat3 operator+(const float& t, const mat3& m)
+const mat3 operator+(const long double& t, const mat3& m)
 {
 	return m + t;
 }
-const mat3 operator-(const float& t, const mat3& m)
+const mat3 operator-(const long double& t, const mat3& m)
 {
 	return m - t;
 }
-const mat3 operator*(const float& t, const mat3& m)
+const mat3 operator*(const long double& t, const mat3& m)
 {
 	return m * t;
 }
-const mat3 operator/(const float& t, const mat3& m)
+const mat3 operator/(const long double& t, const mat3& m)
 {
 	return m / t;
 }
