+++ PSiVR/PSiVR/Geometry/vec3.cpp
@@ -1,19 +1,19 @@
 
 #include "vec3.h"
 
-const vec3 operator+(const float& t, const vec3& v)
+const vec3 operator+(const long double& t, const vec3& v)
 {
 	return v + t;
 }
-const vec3 operator-(const float& t, const vec3& v)
+const vec3 operator-(const long double& t, const vec3& v)
 {
 	return v - t;
 }
-const vec3 operator*(const float& t, const vec3& v)
+const vec3 operator*(const long double& t, const vec3& v)
 {
 	return v * t;
 }
-const vec3 operator/(const float& t, const vec3& v)
+const vec3 operator/(const long double& t, const vec3& v)
 {
 	return v / t;
 }
